test(infinite-scroll): migrate scroll-target
